LPS-116770 Move moveItem shortcuts logic to ShortcutManager
